+++ modules/septima-lang/tests/septima.spec.ts
@@ -285,6 +285,10 @@ describe('septima', () => {
         'sun',
       ])
     })
+    test('arrayness of a value can be tested via Array.isArry()', () => {
+      expect(run(`Array.isArray([1,2,'abc'])`)).toEqual(true)
+      expect(run(`Array.isArray('abc')`)).toEqual(false)
+    })
   })
 
   describe('objects', () => {
@@ -834,6 +838,30 @@ describe('septima', () => {
       expect(run(`Number(undefined)`)).toEqual(NaN)
       expect(run(`Number({})`)).toEqual(NaN)
       expect(run(`Number([])`)).toEqual(NaN)
+    })
+  })
+  describe('console.log', () => {
+    const runLog = (input: string) => {
+      const lines: unknown[] = []
+      const result = Septima.run(input, { onSink: () => undefined, consoleLog: u => lines.push(u) })
+      return { lines, result }
+    }
+    test('prints its input', () => {
+      expect(runLog(`console.log(2*2*2*2)`).lines).toEqual(['16'])
+      expect(runLog(`console.log({a: 1, b: 2, c: ['d', 'e']})`).lines).toEqual(['{"a":1,"b":2,"c":["d","e"]}'])
+    })
+    test('a program can have multiple console.log() calls', () => {
+      expect(runLog(`["red", "green", "blue"].map(at => console.log(at))`).lines).toEqual([
+        '"red"',
+        '"green"',
+        '"blue"',
+      ])
+    })
+    test('returns its input', () => {
+      expect(runLog(`32*console.log(8)`)).toEqual({
+        result: 256,
+        lines: ['8'],
+      })
     })
   })
   test.todo('support file names in locations')