+++ modules/septima-lang/change-log.md
@@ -18,3 +18,8 @@
 - spreading `undefined` in an array is a no-op: `[42, ...undefined, 'poo']` is `[42, 'poo']`
 - spreading `undefined` in an object is a no-op: `{n: 42, ...undefined, p: 'poo'}` is `{n: 42, p: 'poo'}`
 - support `String(x)`, `Boolean(x)`, `Number(x)`
+
+### PR/168
+
+- support `Array.isArray(x)`
+- support `console.log('my message')`